Separate sound into its own thread
