+++ client/engine/screen_manager.py
@@ -1,8 +1,7 @@
-import threading
-from queue import Empty, SimpleQueue
 from typing import TYPE_CHECKING, Any
 
 from client.engine.event_handler import EventHandler
+from client.engine.features.sound.worker import SoundWorker
 from client.engine.general_state.current_screen import CurrentScreen
 from client.engine.general_state.options import Options
 from client.engine.general_state.profile_manager import ProfileManager
@@ -20,43 +19,6 @@ if TYPE_CHECKING:
     from client.engine.primitives.event import Event
 
 
-class StopThread(Exception):
-    """Exception raised to signal a thread to stop processing."""
-
-    pass
-
-
-class SoundWorker(threading.Thread):
-
-    def __init__(self, name, my_queue):
-        super().__init__()
-        self.name = name
-        self.my_queue = my_queue
-        # Event used to signal the thread to stop gracefully
-        self.stop_event = threading.Event()
-        # Log that the worked has started?
-
-    def run(self):
-        """The main execution loop for the thread."""
-        print(f"[{self.name}] Thread started, waiting for events...")
-        while not self.stop_event.is_set():
-            try:
-                # Use a small timeout to allow checking the stop_event
-                event_data = self.my_queue.get(timeout=0.1)
-                # self._process_event(event_data)
-                # self.my_queue.task_done() # Indicate that the task is finished
-            except Empty:
-                # Expected when the queue is empty after the timeout
-                print("No events")
-                continue
-            except StopThread:
-                # Internal exception to cleanly exit the loop
-                break
-            except Exception as e:
-                print(f"Error {e}")
-                break
-
-
 class ScreenManagerFactory:
     @staticmethod
     def create(
@@ -68,26 +30,14 @@ class ScreenManagerFactory:
         QueueManager().initialize(initial_event)
         CurrentScreen().initialize()
         Options().initialize()
-        """
-        sound_queue = SimpleQueue()
 
+        # Initialize sound thread
         sound_thread = SoundWorker(
             name="Sound",
-            my_queue=sound_queue,
+            queue=QueueManager().get("sound"),
         )
 
         sound_thread.start()
-        """
-
-        """
-        The safe termination code for threads is missing
-            while thread_a.is_alive() or thread_b.is_alive():
-                time.sleep(0.2)
-    
-            # The main thread joins the workers to ensure they terminate
-            thread_a.join()
-            thread_b.join()
-        """
 
         return ScreenManager(
             KeyboardInput(),
